refund on purchase cancel and covid stuff

+++ FinalProject/server/src/main/java/org/sertia/server/bl/StreamingTicketController.java
@@ -7,6 +7,7 @@ import org.sertia.contracts.reports.ClientReport;
 import org.sertia.contracts.screening.ticket.request.CancelStreamingTicketRequest;
 import org.sertia.contracts.screening.ticket.request.StreamingPaymentRequest;
 import org.sertia.contracts.screening.ticket.response.StreamingPaymentResponse;
+import org.sertia.server.bl.Services.CreditCardService;
 import org.sertia.server.bl.Services.Reportable;
 import org.sertia.server.dl.DbUtils;
 import org.sertia.server.dl.HibernateSessionFactory;
@@ -15,12 +16,18 @@ import org.sertia.server.dl.classes.Streaming;
 import org.sertia.server.dl.classes.StreamingLink;
 
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 public class StreamingTicketController implements Reportable {
+    private final CreditCardService creditCardService;
+
+    public StreamingTicketController(CreditCardService creditCardService) {
+        this.creditCardService = creditCardService;
+    }
 
     public StreamingPaymentResponse purchaseStreamingTicket(StreamingPaymentRequest request) {
         Streaming streaming = getStreaming(request.movieId).orElse(null);
@@ -67,14 +74,26 @@ public class StreamingTicketController implements Reportable {
     public SertiaBasicResponse cancelStreamingTicket(CancelStreamingTicketRequest request) {
         int streamingTickerId = request.streamingId;
         return DbUtils.getById(StreamingLink.class, streamingTickerId).map(streamingTicket -> {
+            if (getHoursToLinkActivation(streamingTicket) < 1) {
+                return new SertiaBasicResponse(false)
+                        .setFailReason("can't cancel link that starts in less than an hour");
+            }
+
             try (Session session = HibernateSessionFactory.getInstance().openSession()) {
                 session.delete(streamingTicket);
+                creditCardService.refund(streamingTicket.getCustomerPaymentDetails(), streamingTicket.getPaidPrice() / 2);
                 return new SertiaBasicResponse(true);
             } catch (RuntimeException exception) {
-                return new SertiaBasicResponse(false).setFailReason("failed to delete streaming ticket");
+                return new SertiaBasicResponse(false)
+                        .setFailReason("failed to delete streaming ticket");
             }
         }).orElseGet(() ->
-                new SertiaBasicResponse(false).setFailReason("no streaming with id " + streamingTickerId));
+                new SertiaBasicResponse(false)
+                        .setFailReason("no streaming with id " + streamingTickerId));
+    }
+
+    private long getHoursToLinkActivation(StreamingLink streamingLink) {
+        return ChronoUnit.HOURS.between(LocalDateTime.now(), streamingLink.getActivationStart());
     }
 
     @Override
